Home Automation Update
Added basic home automation functionality. This required some changes to existing code, so I adjusted the sentence handler to try and find a known noun so verbs that have different contexts can cause different action (a light can have a "turn on" function to activate it, while something else might have a "turn left" function. Here the verb "turn" is treated diffferently depending on the noun)

A test home automation system was created with the yeelight example (why yeelight instead of something like phillips hue? I have a yeelight, not a phillips hue, and I don't feel like buying one just for testing.) This is mostly just for testing, and the system is not complete.

The updater was also changed to plugman, and at some point I hope to add an "install" verb for installing addons.

The fix for .apc updating was also applied to apples.