Updates and improvements
yeelight home automation is working better. Run yeelight.connect to connect a yeelight smart lightbulb. it will as you to select one from a list and enter a name. for now, use a name that is one word, like "lights" or "light".

Terminal and plugman were also updated to match the current version from apples.