Updates
Wow this is alot.
Completely redesigned the englishActions plugin. Instead of being based on Spacy, it's now based on Stanza. Based on testing with sample sentences, stanza seems to have a higher accuracy with this use case. The design of the verbs has also been changed. Previously, verbs handled most of the sentence translation and error correcting. Now, bandaid functions (see annotations ins englishActions/speech.py) handle most of the error correcting, so verbs only need to interpret the information required for the action. This reduces the amount of work required to add new verbs. This change required modifications to both homeAutomation and haYeelight. A general knowledge module was created to store general information and classes that are common in day to day life, such as colors, people, times, etc. englishTTS was changed to englishTTS-pyttsx3 and englishTTS-tacotron was added as an experimental plugin. The terminal was updated to the newest APPLES version.  English STT was updated to use the newest version of deepspeech, 0.7.1.