Basic Functionality Working
Most of the basic functionality is added back in. Certain components, such as home automation have not been added yet. Plugins were renamed to use snake_case instead of lowerCamelCase. Most functions in plugins were renamed to use the same. The APMs for the old plugins that have been transferred have been removed.

Some work that was in progress in english_actions was finished, specifically relating to how Thing objects are stored. The things dictionary was removed completely, relying on the new ThingStorage and ThingStorageSession (and more importantly, their subclasses!) to handle the list of objects. There is still some buggy behavior and difficulty when ThingReferences get involved, but this should be fixed relatively soon. The main reason for this change is it allows for sentences and such to be processed using different contexts; that is, it allows two requests to be handled simultaneously for two different users without them seeing each other's data. This is pretty important for servers as it allows for data to be processed without worrying about manually clearing out the internal state of the Thing system.

While most of the code was being refactored, the opportunity was taken to fix some formatting and other issues with the code. There is still more work to do, such as add type annotations to all functions, but work is progressing well.